+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Attributor Changelog
 next
 ----
 
+* Fixed bug in `example_mixin` where lazy_attributes were not evaluated.
+* Fixed bug in `Hash` where the class would refuse to load from another `Attributor::Hash` when there were no keys defined and they were seemingly compatible.
+
 2.5.0
 ----